% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_sum.R
\name{dt_sum}
\alias{dt_sum}
\title{Data.table Summen von Variablen}
\usage{
dt_sum(dt, ind_var, dep_var, dep_on, bylist, na.rm = TRUE)
}
\arguments{
\item{dt}{Datatable}

\item{ind_var}{indipendent variable vector index or names}

\item{dep_var}{dependent variable  vector index or names}

\item{dep_on}{dependent on varibale name or index one or per dependent variable}

\item{bylist}{list of group by elements}

\item{na.rm}{handles the na.rm variable at summing}
}
\value{
summarized data.table
}
\description{
creates a sum dependend of variables
}
